+++ processor/src/test/resources/bad_input/com/example/context_ref/BadTypeModel.java
@@ -1,0 +1,14 @@
+package com.example.context_ref;
+
+import org.realityforge.arez.annotations.ArezComponent;
+import org.realityforge.arez.annotations.ContextRef;
+
+@ArezComponent( allowEmpty = true )
+class BadTypeModel
+{
+  @ContextRef
+  String getContext()
+  {
+    throw new IllegalStateException();
+  }
+}