+++ processor/src/test/resources/expected/com/example/context_ref/Arez_SimpleComponent.java
@@ -1,0 +1,75 @@
+package com.example.context_ref;
+
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import org.realityforge.arez.Arez;
+import org.realityforge.arez.ArezContext;
+import org.realityforge.arez.Disposable;
+
+@Generated("org.realityforge.arez.processor.ArezProcessor")
+final class Arez_SimpleComponent extends SimpleComponent implements Disposable {
+  private static volatile long $$arez$$_nextId;
+
+  private final long $$arez$$_id;
+
+  private boolean $$arez$$_disposed;
+
+  @Nonnull
+  private final ArezContext $$arez$$_context;
+
+  Arez_SimpleComponent() {
+    super();
+    this.$$arez$$_context = Arez.context();
+    this.$$arez$$_id = $$arez$$_nextId++;
+  }
+
+  final ArezContext getContext() {
+    return $$arez$$_context;
+  }
+
+  final long $$arez$$_id() {
+    return $$arez$$_id;
+  }
+
+  String $$arez$$_name() {
+    return "SimpleComponent." + $$arez$$_id();
+  }
+
+  @Override
+  public boolean isDisposed() {
+    return $$arez$$_disposed;
+  }
+
+  @Override
+  public void dispose() {
+    if ( !isDisposed() ) {
+      $$arez$$_disposed = true;
+    }
+  }
+
+  @Override
+  public final int hashCode() {
+    return Long.hashCode( $$arez$$_id() );
+  }
+
+  @Override
+  public final boolean equals(final Object o) {
+    if ( this == o ) {
+      return true;
+    } else if ( null == o || !(o instanceof Arez_SimpleComponent) ) {
+      return false;
+    } else {
+      final Arez_SimpleComponent that = (Arez_SimpleComponent) o;;
+      return $$arez$$_id() == that.$$arez$$_id();
+    }
+  }
+
+  @Override
+  public final String toString() {
+    if ( Arez.areNamesEnabled() ) {
+      return "ArezComponent[" + $$arez$$_name() + "]";
+    } else {
+      return super.toString();
+    }
+  }
+}