+++ processor/src/test/resources/input/com/example/context_ref/SimpleComponent.java
@@ -1,0 +1,15 @@
+package com.example.context_ref;
+
+import org.realityforge.arez.ArezContext;
+import org.realityforge.arez.annotations.ArezComponent;
+import org.realityforge.arez.annotations.ContextRef;
+
+@ArezComponent( allowEmpty = true )
+class SimpleComponent
+{
+  @ContextRef
+  ArezContext getContext()
+  {
+    throw new IllegalStateException();
+  }
+}